Rename Hero to LandingPage. Place Header inside LandingPage
