+++ src/components/LandingPage.css
@@ -1,5 +1,15 @@
+.landingPage {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
 .hero {
-  padding: 260px 100px;
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  padding: 72px 100px;
   text-align: center;
   background-color: var(--primary-color-dark);
 }
@@ -18,18 +28,18 @@
   font-weight: 700;
   color: var(--primary-color-light);
   font-family: "Inter", sans-serif;
+  
 }
-/* .hero  {
-  font-size: 24px;
-  margin-bottom: 20px;
-} */
-
 .hero .see-more {
   padding: 24px 0px;
   display: flex;
   flex: stretch;
   justify-content: flex-end;
+  position:absolute;
+  bottom: 24px;
+  right: 100px;
   width: 100%;
+
 }
 .hero .see-more button{
   all: unset;
@@ -44,7 +54,6 @@
   animation-name: bounce;
   animation-duration: 1.5s;
   animation-iteration-count: infinite;
-  /* animation-timing-function: cubic-bezier(0.95, 0.05, 0.795, 0.035); */
   animation-direction: alternate;
 }
 .hero .see-more button:hover {
